+++ postbound/db/postgres.py
@@ -3001,7 +3001,7 @@ def start(pgdata: str | Path = "") -> None:
     if os.system("which pg_ctl") != 0:
         raise ValueError("Cannot start Postgres server: pg_ctl is not on PATH")
 
-    pgdata = pgdata or os.environ["PGDATA"]
+    pgdata = pgdata or os.environ.get("PGDATA", "")
     pgdata = Path(pgdata).expanduser()
     if not pgdata:
         raise ValueError(
@@ -3023,7 +3023,7 @@ def stop(pgdata: str | Path = "", *, raise_on_error: bool = False) -> None:
     if os.system("which pg_ctl") != 0:
         raise ValueError("Cannot stop Postgres server: pg_ctl is not on PATH")
 
-    pgdata = pgdata or os.environ["PGDATA"]
+    pgdata = pgdata or os.environ.get("PGDATA", "")
     pgdata = Path(pgdata).expanduser()
     if not pgdata:
         raise ValueError(
@@ -3031,6 +3031,26 @@ def stop(pgdata: str | Path = "", *, raise_on_error: bool = False) -> None:
         )
 
     subprocess.run(["pg_ctl", "-D", pgdata, "stop"], check=raise_on_error)
+
+
+def is_running(pgdata: str | Path = "") -> bool:
+    """Checks, whether a local Postgres server is currently running.
+
+    This function assumes that *pg_ctl* is available on the system PATH. A data directory can be supplied to check whether
+    a server is running for the specific database. If *pgdata* is not supplied, the *PGDATA* environment variable is used as
+    a fallback.
+    """
+    if os.system("which pg_ctl") != 0:
+        raise ValueError("Cannot start Postgres server: pg_ctl is not on PATH")
+
+    cmd = ["pg_ctl"]
+    pgdata = pgdata or os.environ.get("PGDATA", "")
+    if pgdata:
+        cmd.extend(["-D", pgdata])
+    cmd.append("status")
+
+    res = subprocess.run(cmd)
+    return res.returncode == 0
 
 
 def _parallel_query_initializer(